test: split PDF visual composite recovery coverage
